Latest and greatest from PhiPro. More desync corner cases dealt with.

+++ src/main/java/nessiesson/newlight/IChunk.java
@@ -4,9 +4,9 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.EnumSkyBlock;
 
 public interface IChunk {
-	short[] getNeighborLightChecks();
+	int[] getNeighborLightChecks();
 
-	void setNeighborLightChecks(short[] in);
+	void setNeighborLightChecks(int[] in);
 
 	short getPendingNeighborLightInits();
